-- Get Access token with rxJava

+++ app/src/main/java/br/com/eduardofilho/ifood_mobile_test/model/User.kt
@@ -2,7 +2,7 @@ package br.com.eduardofilho.ifood_mobile_test.model
 
 import com.google.gson.annotations.SerializedName
 
-data class User(@SerializedName("id") val id : Int,
+data class User(@SerializedName("id") val id : Long,
                 @SerializedName("name") val name : String,
                 @SerializedName("screen_name") val screenName : String,
                 @SerializedName("url") val url : String)
